pyproject.toml
[tool.poetry]
name = "salt_lake_satellite"
version = "0.1.0"
description = ""
authors = ["Rachel Morrison <rmorrisonca@gmail.com>"]
readme = "README.md"
packages = [{include = "salt_lake_satellite"}]

[tool.poetry.dependencies]
python = "^3.11"
geopandas = "^0.14.3"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
numpy = "^1.26.3"
rasterio = "^1.3.9"
rioxarray = "^0.15.1"
earthpy = "^0.9.4"
shapely = "^2.0.2"
PyYAML = "^6.0.1"
torchgeo = "^0.5.1"
scikit-learn = "^1.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
